trying basic on gpu

+++ parallel-scan/kernels.cu
@@ -8,9 +8,9 @@
 #define LOG_MEM_BANKS 5
 #define CONFLICT_FREE_OFFSET(n) ((n) >> LOG_MEM_BANKS)
 
-__global__ void prescan_arbitrary(int *output, int *input, int n, int powerOfTwo)
+__global__ void prescan_arbitrary(double *output, double *input, int n, int powerOfTwo)
 {
-	extern __shared__ int temp[];// allocated on invocation
+	extern __shared__ double temp[];// allocated on invocation
 	int threadID = threadIdx.x;
 
 	int ai = threadID;
@@ -73,8 +73,8 @@ __global__ void prescan_arbitrary(int *output, int *input, int n, int powerOfTwo
 	}
 }
 
-__global__ void prescan_arbitrary_unoptimized(int *output, int *input, int n, int powerOfTwo) {
-	extern __shared__ int temp[];// allocated on invocation
+__global__ void prescan_arbitrary_unoptimized(double *output, double *input, int n, int powerOfTwo) {
+	extern __shared__ double temp[];// allocated on invocation
 	int threadID = threadIdx.x;
 
 	if (threadID < n) {
@@ -124,8 +124,8 @@ __global__ void prescan_arbitrary_unoptimized(int *output, int *input, int n, in
 }
 
 
-__global__ void prescan_large(int *output, int *input, int n, int *sums) {
-	extern __shared__ int temp[];
+__global__ void prescan_large(double *output, double *input, int n, int *sums) {
+	extern __shared__ double temp[];
 
 	int blockID = blockIdx.x;
 	int threadID = threadIdx.x;
@@ -183,12 +183,12 @@ __global__ void prescan_large(int *output, int *input, int n, int *sums) {
 	output[blockOffset + bi] = temp[bi + bankOffsetB];
 }
 
-__global__ void prescan_large_unoptimized(int *output, int *input, int n, int *sums) {
+__global__ void prescan_large_unoptimized(double *output, double *input, int n, int *sums) {
 	int blockID = blockIdx.x;
 	int threadID = threadIdx.x;
 	int blockOffset = blockID * n;
 
-	extern __shared__ int temp[];
+	extern __shared__ double temp[];
 	temp[2 * threadID] = input[blockOffset + (2 * threadID)];
 	temp[2 * threadID + 1] = input[blockOffset + (2 * threadID) + 1];
 
@@ -232,7 +232,7 @@ __global__ void prescan_large_unoptimized(int *output, int *input, int n, int *s
 }
 
 
-__global__ void add(int *output, int length, int *n) {
+__global__ void add(double *output, int length, double *n) {
 	int blockID = blockIdx.x;
 	int threadID = threadIdx.x;
 	int blockOffset = blockID * length;
@@ -240,7 +240,7 @@ __global__ void add(int *output, int length, int *n) {
 	output[blockOffset + threadID] += n[blockID];
 }
 
-__global__ void add(int *output, int length, int *n1, int *n2) {
+__global__ void add(double *output, int length, double *n1, double *n2) {
 	int blockID = blockIdx.x;
 	int threadID = threadIdx.x;
 	int blockOffset = blockID * length;
