+++ parallel-scan/kernels.cuh
@@ -1,8 +1,8 @@
-__global__ void prescan_arbitrary(int *g_odata, int *g_idata, int n, int powerOfTwo);
-__global__ void prescan_arbitrary_unoptimized(int *g_odata, int *g_idata, int n, int powerOfTwo);
+__global__ void prescan_arbitrary(double *g_odata, double *g_idata, int n, int powerOfTwo);
+__global__ void prescan_arbitrary_unoptimized(double *g_odata, double *g_idata, int n, int powerOfTwo);
 
-__global__ void prescan_large(int *g_odata, int *g_idata, int n, int* sums);
-__global__ void prescan_large_unoptimized(int *output, int *input, int n, int *sums);
+__global__ void prescan_large(double *g_odata, double *g_idata, int n, int* sums);
+__global__ void prescan_large_unoptimized(double *output, double *input, int n, int *sums);
 
-__global__ void add(int *output, int length, int *n1);
-__global__ void add(int *output, int length, int *n1, int *n2);
+__global__ void add(double *output, int length, double *n1);
+__global__ void add(double *output, int length, double *n1, double *n2);
